Move `getFileNameFromPath` to `FileUtils` for consistency.
